+++ src/expansion_boards/mod.rs
@@ -7,5 +7,5 @@ pub use xb200_filter::*;
 mod xb200_path;
 pub use xb200_path::*;
 
-pub mod xb200_gpio;
 pub(crate) mod xb_gpio;
+pub mod xb_gpio_impls;
